feat: cli new approach
