+++ packages/core/src/experimental/searchServerSideFunctions/index.ts
@@ -1,0 +1,2 @@
+export * from './getServerSideProps'
+export * from './getStaticProps'
